Improved waiting for the tailer to start
It was theoretically possible for some tests to write into a file before
the scheduled thread executor within the Tailer could kick in. This, in
turn, would result in no messages being received for cases where the
tailer was set to read files from the end. The content had already been
written before the tailing actually, internally started.